rename file & fixing path
