+++ src/modules/lookup/classes/filters/menu/index.ts
@@ -1,0 +1,2 @@
+export * from './IMenuGetAllFilter';
+export * from './IMenuListFilter';
