+++ src/modules/lookup/classes/queries/menu/IMenuGetAllRequest.ts
@@ -1,4 +1,4 @@
-import { IMenuGetAllFilter } from '@lookup/classes/filters';
+import { IMenuGetAllFilter } from '@lookup/classes/filters/menu';
 
 export interface IMenuGetAllRequest {
   readonly filter: IMenuGetAllFilter | undefined;
