+++ src/modules/lookup/classes/queries/menu/IMenuListRequest.ts
@@ -1,4 +1,4 @@
-import { IMenuListFilter } from '@lookup/classes/filters';
+import { IMenuListFilter } from '@lookup/classes/filters/menu';
 
 export interface IMenuListRequest {
   readonly filter: IMenuListFilter | undefined;
