+++ src/modules/lookup/classes/queries/menu/index.ts
@@ -1,0 +1,3 @@
+export * from './IMenuGetAllRequest';
+export * from './IMenuGetByIdRequest';
+export * from './IMenuListRequest';
