+++ src/modules/lookup/classes/response/menu/index.ts
@@ -1,0 +1,6 @@
+export * from './IMenu';
+export * from './IMenuDetail';
+export * from './IMenuList';
+export * from './IMenuParent';
+export * from './IMenuParentDetail';
+export * from './IMenuParentList';
