+++ src/modules/lookup/store/reducers/menu/lookupMenuReducer.ts
@@ -1,0 +1,9 @@
+import { lookupMenuGetAllReducer } from './lookupMenuGetAllReducer';
+import { lookupMenuGetByIdReducer } from './lookupMenuGetByIdReducer';
+import { lookupMenuGetListReducer } from './lookupMenuGetListReducer';
+
+export const lookupMenuReducers = {
+  lookupMenuGetAll: lookupMenuGetAllReducer,
+  lookupMenuGetList: lookupMenuGetListReducer,
+  lookupMenuGetById: lookupMenuGetByIdReducer,
+};
